split BasicUtils in Shape, Math, Table Utils and Constants
