+++ src/util/Constants.cpp
@@ -1,0 +1,19 @@
+
+#include "Constants.h"
+
+//-----------------------------------------------------------------------
+//   constants
+//-----------------------------------------------------------------------
+
+//                                  1D  2D  3D
+const int    STRAIN_COUNTS[4] = { 0, 1, 4, 6 };
+
+// Other constants
+
+const double PI               = 3.14159265;
+const double PI3              = 0.3333333333333333 * PI;
+const double ONE_THIRD        = 0.3333333333333333;
+const double TWO_THIRD        = 0.6666666666666667;
+const double EPS              = 1.e-16;
+const double RAD_120          = 0.6666666666666667 * PI;
+const double RAD_240          = 2.0 * RAD_120;
