+++ src/util/Constants.h
@@ -1,0 +1,19 @@
+
+#ifndef CONSTANTS_H
+#define CONSTANTS_H
+
+//-----------------------------------------------------------------------
+//   constants
+//-----------------------------------------------------------------------
+
+extern const int      STRAIN_COUNTS[4];
+extern const double   PI;
+extern const double   PI3;
+extern const double   ONE_THIRD;
+extern const double   TWO_THIRD;
+extern const double   EPS;
+extern const double   RAD_120;
+extern const double   RAD_240;
+
+#endif
+
